Ultimos ajustes Case C#

+++ CooperativaCreditoAPI/Repositories/IRepository.cs
@@ -1,6 +1,6 @@
 public interface IRepository<T> where T : class
 {
-    T? GetById(int id);
+    T? GetById(long id);
     IEnumerable<T> GetAll();
     void Add(T entity);
     void Update(T entity);
